Tests/LibWeb/Text/expected/navigation/history-pushstate.txt

history object length has changed by 1
